Allow circular check to be disabled
The circular check is a pretty expensive operation in the case of large graphs,
so allow it to be disabled via a keyword argument when adding DAG edges. This is
useful for cases when the user is absolutely sure that there are no cycles in
the graph, e.g. when importing a properly sanitized graph.